@@ -1,6 +1,8 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+
 import "./App.css";
+
 import NavMenu from "./components/NavMenu";
 import About from "./pages/About";
 import Home from "./pages/Home";
@@ -16,10 +18,10 @@ function App() {
         <NavMenu />
         <ScrollToTop />
         <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="/about" element={<About />} />
-          <Route path="/projects" element={<Projects />} />
-          <Route path="/contact" element={<Contact />} />
+          <Route path="/" Component={Home} />
+          <Route path="/about" Component={About} />
+          <Route path="/projects" Component={Projects} />
+          <Route path="/contact" Component={Contact} />
         </Routes>
         <Footer />
       </Router>
