@@ -1,9 +1,10 @@
 import React from "react";
+import styled from "styled-components";
+
 import Button from "./Button";
 import PText from "./PText";
 import SectionTitle from "./SectionTitle";
 import AboutImage from "../assets/images/about-me-img-2.jpg";
-import styled from "styled-components";
 
 const AboutSectionStyle = styled.div`
   padding: 10rem 0;
